refactor(auth): delegate all authentication to core
Authentication of multiple users on a connected MongoClient is
prohibited, so we should no longer explicitly auth in native. All
authentication is handled by the underlaying pool, so all we
should be doing is generating credentials and letting the existing
machinery handle the rest.

NODE-1859
