+++ lib/mongo_client.js
@@ -172,7 +172,7 @@ MongoClient.prototype.connect = function(callback) {
 
 MongoClient.prototype.logout = deprecate(function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
-  callback(null, true);
+  if (typeof callback === 'function') callback(null, true);
 }, 'Multiple authentication is prohibited on a connected client, please only authenticate once per MongoClient');
 
 /**
