+++ lib/operations/mongo_client_ops.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const authenticate = require('../authenticate').authenticate;
-const generateCredentials = require('../authenticate').generateCredentials;
 const deprecate = require('util').deprecate;
 const Logger = require('mongodb-core').Logger;
 const MongoError = require('mongodb-core').MongoError;
@@ -12,6 +10,7 @@ const ReplSet = require('../topologies/replset');
 const Server = require('../topologies/server');
 const ServerSessionPool = require('mongodb-core').Sessions.ServerSessionPool;
 const NativeTopology = require('../topologies/native_topology');
+const MongoCredentials = require('mongodb-core').MongoCredentials;
 
 let client;
 function loadClient() {
@@ -181,7 +180,7 @@ function connect(mongoClient, url, options, callback) {
     throw new Error('no callback function provided');
   }
 
-  // Get a logger for MongoClient
+  let didRequestAuthentication = false;
   const logger = Logger('MongoClient', options);
 
   // Did we pass in a Server/ReplSet/Mongos
@@ -220,6 +219,7 @@ function connect(mongoClient, url, options, callback) {
 
     if (_finalOptions.auth && !_finalOptions.credentials) {
       try {
+        didRequestAuthentication = true;
         _finalOptions.credentials = generateCredentials(
           mongoClient,
           _finalOptions.auth.user,
@@ -232,35 +232,16 @@ function connect(mongoClient, url, options, callback) {
     }
 
     if (_finalOptions.useUnifiedTopology) {
-      return createTopology(
-        mongoClient,
-        'unified',
-        _finalOptions,
-        connectHandler(mongoClient, _finalOptions, connectCallback)
-      );
+      return createTopology(mongoClient, 'unified', _finalOptions, connectCallback);
     }
 
     // Do we have a replicaset then skip discovery and go straight to connectivity
     if (_finalOptions.replicaSet || _finalOptions.rs_name) {
-      return createTopology(
-        mongoClient,
-        'replicaset',
-        _finalOptions,
-        connectHandler(mongoClient, _finalOptions, connectCallback)
-      );
+      return createTopology(mongoClient, 'replicaset', _finalOptions, connectCallback);
     } else if (object.servers.length > 1) {
-      return createTopology(
-        mongoClient,
-        'mongos',
-        _finalOptions,
-        connectHandler(mongoClient, _finalOptions, connectCallback)
-      );
+      return createTopology(mongoClient, 'mongos', _finalOptions, connectCallback);
     } else {
-      return createServer(
-        mongoClient,
-        _finalOptions,
-        connectHandler(mongoClient, _finalOptions, connectCallback)
-      );
+      return createServer(mongoClient, _finalOptions, connectCallback);
     }
   });
 
@@ -275,33 +256,13 @@ function connect(mongoClient, url, options, callback) {
       return callback(new MongoError(warningMessage));
     }
 
+    if (didRequestAuthentication) {
+      mongoClient.emit('authenticated', null, true);
+    }
+
     // Return the error and db instance
     callback(err, topology);
   }
-}
-
-function connectHandler(client, options, callback) {
-  return (err, topology) => {
-    if (err) {
-      return handleConnectCallback(err, topology, callback);
-    }
-
-    // No authentication just reconnect
-    if (!options.credentials) {
-      return handleConnectCallback(err, topology, callback);
-    }
-
-    // Authenticate
-    authenticate(client, options.credentials, (err, success) => {
-      if (success) {
-        handleConnectCallback(null, topology, callback);
-      } else {
-        if (topology) topology.close();
-        const authError = err ? err : new Error('Could not authenticate user ' + options.auth[0]);
-        handleConnectCallback(authError, topology, callback);
-      }
-    });
-  };
 }
 
 /**
@@ -345,7 +306,6 @@ function connectWithUrl(mongoClient, url, options, connectCallback) {
   }
 
   const isDoingAuth = finalOptions.user || finalOptions.password || finalOptions.authMechanism;
-
   if (isDoingAuth && !finalOptions.credentials) {
     try {
       finalOptions.credentials = generateCredentials(
@@ -359,21 +319,7 @@ function connectWithUrl(mongoClient, url, options, connectCallback) {
     }
   }
 
-  // Connect
-  return url.connect(
-    finalOptions,
-    connectHandler(mongoClient, finalOptions, (err, topology) => {
-      if (err) return connectCallback(err, topology);
-      if (isDoingAuth) {
-        return authenticate(mongoClient, finalOptions.credentials, err => {
-          if (err) return connectCallback(err, topology);
-          connectCallback(err, topology);
-        });
-      }
-
-      connectCallback(err, topology);
-    })
-  );
+  return url.connect(finalOptions, connectCallback);
 }
 
 function createListener(mongoClient, event) {
@@ -504,17 +450,6 @@ function createUnifiedOptions(finalOptions, options) {
   }
 
   return finalOptions;
-}
-
-function handleConnectCallback(err, topology, callback) {
-  return process.nextTick(() => {
-    try {
-      callback(err, topology);
-    } catch (err) {
-      if (topology) topology.close();
-      throw err;
-    }
-  });
 }
 
 function legacyTransformUrlOptions(object) {
@@ -671,6 +606,61 @@ function validOptions(options) {
       );
     }
   }
+}
+
+const VALID_AUTH_MECHANISMS = new Set([
+  'DEFAULT',
+  'MONGODB-CR',
+  'PLAIN',
+  'MONGODB-X509',
+  'SCRAM-SHA-1',
+  'SCRAM-SHA-256',
+  'GSSAPI'
+]);
+
+const AUTH_MECHANISM_INTERNAL_MAP = {
+  DEFAULT: 'default',
+  'MONGODB-CR': 'mongocr',
+  PLAIN: 'plain',
+  'MONGODB-X509': 'x509',
+  'SCRAM-SHA-1': 'scram-sha-1',
+  'SCRAM-SHA-256': 'scram-sha-256'
+};
+
+function generateCredentials(client, username, password, options) {
+  options = Object.assign({}, options);
+
+  // the default db to authenticate against is 'self'
+  // if authententicate is called from a retry context, it may be another one, like admin
+  const source = options.authSource || options.authdb || options.dbName;
+
+  // authMechanism
+  const authMechanismRaw = options.authMechanism || 'DEFAULT';
+  const authMechanism = authMechanismRaw.toUpperCase();
+
+  if (!VALID_AUTH_MECHANISMS.has(authMechanism)) {
+    throw MongoError.create({
+      message: `authentication mechanism ${authMechanismRaw} not supported', options.authMechanism`,
+      driver: true
+    });
+  }
+
+  if (authMechanism === 'GSSAPI') {
+    return new MongoCredentials({
+      mechanism: process.platform === 'win32' ? 'sspi' : 'gssapi',
+      mechanismProperties: options,
+      source,
+      username,
+      password
+    });
+  }
+
+  return new MongoCredentials({
+    mechanism: AUTH_MECHANISM_INTERNAL_MAP[authMechanism],
+    source,
+    username,
+    password
+  });
 }
 
 module.exports = { connectOp, validOptions };
