+++ lib/topologies/topology_base.js
@@ -370,11 +370,6 @@ class TopologyBase extends EventEmitter {
     return this.s.coreTopology.unref();
   }
 
-  auth() {
-    var args = Array.prototype.slice.call(arguments, 0);
-    this.s.coreTopology.auth.apply(this.s.coreTopology, args);
-  }
-
   /**
    * All raw connections
    * @method
